moving tests to test_commons; adding test for no warning
